Split the code of drawing battery status
