+++ main/gui/battery_status.hpp
@@ -1,0 +1,6 @@
+#ifndef M5CORE2_TIME_BOX_GUI_BATTERY_STATUS_HPP_
+#define M5CORE2_TIME_BOX_GUI_BATTERY_STATUS_HPP_
+
+void drawBatteryPercentage();
+
+#endif // M5CORE2_TIME_BOX_GUI_BATTERY_STATUS_HPP_
